+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/plugin/jei/wrapper/JEIRecipeWrapperCampfire.java
@@ -2,11 +2,9 @@ package com.codetaylor.mc.pyrotech.modules.tech.basic.plugin.jei.wrapper;
 
 import com.codetaylor.mc.athenaeum.util.StringHelper;
 import com.codetaylor.mc.pyrotech.library.spi.plugin.jei.IPyrotechRecipeWrapper;
-import com.codetaylor.mc.pyrotech.modules.tech.basic.ModuleTechBasicConfig;
 import com.codetaylor.mc.pyrotech.modules.tech.basic.recipe.CampfireRecipe;
 import mezz.jei.api.ingredients.IIngredients;
 import mezz.jei.api.ingredients.VanillaTypes;
-import mezz.jei.api.recipe.IRecipeWrapper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.Ingredient;
@@ -21,27 +19,27 @@ import java.util.List;
 public class JEIRecipeWrapperCampfire
     implements IPyrotechRecipeWrapper {
 
-  private static final String TIME_STRING = StringHelper.ticksToHMS(ModuleTechBasicConfig.CAMPFIRE.COOK_TIME_TICKS);
-
   private final ResourceLocation registryName;
+  private final String timeString;
   private final List<List<ItemStack>> inputs;
   private final ItemStack output;
 
   public JEIRecipeWrapperCampfire(CampfireRecipe recipe) {
 
-    this(recipe.getInput(), recipe.getOutput(), recipe.getRegistryName());
+    this(recipe.getInput(), recipe.getOutput(), recipe.getRegistryName(), recipe.getTicks());
   }
 
-  public JEIRecipeWrapperCampfire(Ingredient input, ItemStack output) {
+  public JEIRecipeWrapperCampfire(Ingredient input, ItemStack output, int ticks) {
 
-    this(input, output, null);
+    this(input, output, null, ticks);
   }
 
-  private JEIRecipeWrapperCampfire(Ingredient input, ItemStack output, ResourceLocation registryName) {
+  private JEIRecipeWrapperCampfire(Ingredient input, ItemStack output, ResourceLocation registryName, int ticks) {
 
     this.inputs = Collections.singletonList(Arrays.asList(input.getMatchingStacks()));
     this.output = output;
     this.registryName = registryName;
+    this.timeString = StringHelper.ticksToHMS(ticks);
   }
 
   @Override
@@ -54,8 +52,8 @@ public class JEIRecipeWrapperCampfire
   @Override
   public void drawInfo(Minecraft minecraft, int recipeWidth, int recipeHeight, int mouseX, int mouseY) {
 
-    int stringWidth = minecraft.fontRenderer.getStringWidth(TIME_STRING);
-    minecraft.fontRenderer.drawString(TIME_STRING, 36 - stringWidth / 2, 22, Color.DARK_GRAY.getRGB());
+    int stringWidth = minecraft.fontRenderer.getStringWidth(this.timeString);
+    minecraft.fontRenderer.drawString(this.timeString, 36 - stringWidth / 2, 22, Color.DARK_GRAY.getRGB());
   }
 
   @Override